move energy data logging into data module
